+++ src/common/homography.h
@@ -130,12 +130,12 @@ extern "C" {
 
 matd_t *homography_compute(zarray_t *correspondences, int flags);
 
-//void homography_project(const matd_t *H, double x, double y, double *ox, double *oy);
-static inline void homography_project(const matd_t *H, double x, double y, double *ox, double *oy)
+//void homography_project(const matd_t *H, float x, float y, float *ox, float *oy);
+static inline void homography_project(const matd_t *H, float x, float y, float *ox, float *oy)
 {
-    double xx = MATD_EL(H, 0, 0)*x + MATD_EL(H, 0, 1)*y + MATD_EL(H, 0, 2);
-    double yy = MATD_EL(H, 1, 0)*x + MATD_EL(H, 1, 1)*y + MATD_EL(H, 1, 2);
-    double zz = MATD_EL(H, 2, 0)*x + MATD_EL(H, 2, 1)*y + MATD_EL(H, 2, 2);
+    float xx = MATD_EL(H, 0, 0)*x + MATD_EL(H, 0, 1)*y + MATD_EL(H, 0, 2);
+    float yy = MATD_EL(H, 1, 0)*x + MATD_EL(H, 1, 1)*y + MATD_EL(H, 1, 2);
+    float zz = MATD_EL(H, 2, 0)*x + MATD_EL(H, 2, 1)*y + MATD_EL(H, 2, 2);
 
     *ox = xx / zz;
     *oy = yy / zz;
@@ -166,7 +166,7 @@ static inline void homography_project(const matd_t *H, double x, double y, doubl
 // R20 = H20
 // R21 = H21
 // TZ  = H22
-matd_t *homography_to_pose(const matd_t *H, double fx, double fy, double cx, double cy);
+matd_t *homography_to_pose(const matd_t *H, float fx, float fy, float cx, float cy);
 
 // Similar to above
 // Recover the model view matrix assuming that the projection matrix is:
@@ -176,7 +176,7 @@ matd_t *homography_to_pose(const matd_t *H, double fx, double fy, double cx, dou
 // [ 0  0  C  D ]
 // [ 0  0 -1  0 ]
 
-matd_t *homography_to_model_view(const matd_t *H, double F, double G, double A, double B);
+matd_t *homography_to_model_view(const matd_t *H, float F, float G, float A, float B);
 
 #ifdef __cplusplus
 }
