+++ src/common/image_u8.h
@@ -68,7 +68,7 @@ void image_u8_fill_line_max(image_u8_t *im, const image_u8_lut_t *lut, const flo
 void image_u8_clear(image_u8_t *im);
 void image_u8_darken(image_u8_t *im);
 void image_u8_convolve_2D(image_u8_t *im, const uint8_t *k, int ksz);
-void image_u8_gaussian_blur(image_u8_t *im, double sigma, int k);
+void image_u8_gaussian_blur(image_u8_t *im, float sigma, int k);
 
 // 1.5, 2, 3, 4, ... supported
 image_u8_t *image_u8_decimate(image_u8_t *im, float factor);
@@ -83,7 +83,7 @@ int image_u8_write_pnm(const image_u8_t *im, const char *path);
 // sense", i.e., if Y were up. When input values are unavailable, the
 // value 'pad' is inserted instead. The geometric center of the output
 // image corresponds to the geometric center of the input image.
-image_u8_t *image_u8_rotate(const image_u8_t *in, double rad, uint8_t pad);
+image_u8_t *image_u8_rotate(const image_u8_t *in, float rad, uint8_t pad);
 
 #ifdef __cplusplus
 }
